# sim-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é a API para o projeto de banco de dados, permitindo a gestão de alunos, cursos e o registro de interesse em turmas.",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/interesses": {
            "get": {
                "description": "Retorna uma lista de todos os interesses registrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interesses"
                ],
                "summary": "Lista todos os registros de interesse",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.RegistroInteresse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Registra o interesse de um aluno em uma turma",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interesses"
                ],
                "summary": "Cria um novo registro de interesse",
                "parameters": [
                    {
                        "description": "Dados do Interesse",
                        "name": "interesse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RegistroInteresse"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.RegistroInteresse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/interesses/{id}": {
            "get": {
                "description": "Retorna os detalhes de um único registro de interesse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interesses"
                ],
                "summary": "Busca um registro de interesse pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Registro de Interesse",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.RegistroInteresse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza a prioridade ou status de um registro de interesse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interesses"
                ],
                "summary": "Atualiza um registro de interesse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Registro de Interesse",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para Atualizar",
                        "name": "interesse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RegistroInteresse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.RegistroInteresse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove o interesse de um aluno em uma turma",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interesses"
                ],
                "summary": "Apaga um registro de interesse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do Registro de Interesse",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/meus-interesses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna uma lista de todos os interesses registrados pelo aluno que está autenticado via token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interesses"
                ],
                "summary": "Lista os Interesses do Aluno Logado",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.RegistroInteresse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.RegistroInteresse": {
            "type": "object",
            "properties": {
                "data_registro_interesse": {
                    "type": "string"
                },
                "id_registro_interesse": {
                    "type": "integer"
                },
                "id_turma": {
                    "type": "integer"
                },
                "matricula_aluno": {
                    "type": "string"
                },
                "prioridade_interesse": {
                    "type": "integer"
                },
                "status_interesse": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer \" seguido do seu token JWT. Exemplo: \"Bearer eyJhbGciOi...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}